contrib/common: extract _heartbeat_decorator for cross-plugin use
